fixed gridpane spaces bug
replaced imageview with css background image

+++ src/de/bricked/game/board/Board.java
@@ -3,7 +3,7 @@ package de.bricked.game.board;
 import java.util.ArrayList;
 
 import de.bricked.game.bricks.Brick;
-import de.bricked.game.config.BrickType;
+import de.bricked.game.bricks.BrickType;
 import de.bricked.game.levels.Level;
 import de.bricked.game.levels.LevelPack;
 import de.bricked.game.levels.LevelPackReader;
@@ -12,8 +12,8 @@ public class Board
 {
 	// first = row, second = column
 	private ArrayList<ArrayList<Brick>> bricks;
-	public static final int WIDTH = 18;
-	public static final int HEIGHT = 26;
+	public static final double WIDTH = 18.0;
+	public static final double HEIGHT = 26.0;
 
 	public Board(Level level)
 	{
@@ -141,12 +141,12 @@ public class Board
 		this.bricks = bricks;
 	}
 
-	public int getWidth()
+	public double getWidth()
 	{
 		return WIDTH;
 	}
 
-	public int getHeight()
+	public double getHeight()
 	{
 		return HEIGHT;
 	}
